Move rule examples out of rule file
